+++ src/ball.rs
@@ -1,37 +1,55 @@
-// Copyright © 2022 Rouven Spreckels <rs@qu1x.dev>
+// Copyright © 2022-2024 Rouven Spreckels <rs@qu1x.dev>
 //
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use super::Enclosing;
-use nalgebra::{distance_squared, Point, RealField, SMatrix, SVector};
+use nalgebra::{
+	base::allocator::Allocator, DefaultAllocator, DimNameAdd, DimNameSum, OMatrix, OPoint, OVector,
+	RealField, U1,
+};
 
 /// Ball over real field `R` of dimension `D` with center and radius squared.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct Ball<R: RealField, const D: usize> {
+#[derive(Debug, Clone, PartialEq)]
+pub struct Ball<R: RealField, D: DimNameAdd<U1>>
+where
+	DefaultAllocator: Allocator<R, D>,
+{
 	/// Ball's center.
-	pub center: Point<R, D>,
+	pub center: OPoint<R, D>,
 	/// Ball's radius squared.
 	pub radius_squared: R,
 }
 
-impl<R: RealField, const D: usize> Enclosing<R, D> for Ball<R, D> {
+impl<R: RealField + Copy, D: DimNameAdd<U1>> Copy for Ball<R, D>
+where
+	OPoint<R, D>: Copy,
+	DefaultAllocator: Allocator<R, D>,
+{
+}
+
+impl<R: RealField, D: DimNameAdd<U1>> Enclosing<R, D> for Ball<R, D>
+where
+	DefaultAllocator:
+		Allocator<R, D> + Allocator<R, D, D> + Allocator<OPoint<R, D>, DimNameSum<D, U1>>,
+	<DefaultAllocator as Allocator<OPoint<R, D>, DimNameSum<D, U1>>>::Buffer: Default,
+{
 	#[inline]
-	fn contains(&self, point: &Point<R, D>) -> bool {
-		distance_squared(&self.center, point) <= self.radius_squared
+	fn contains(&self, point: &OPoint<R, D>) -> bool {
+		(point - &self.center).norm_squared() <= self.radius_squared
 	}
-	fn with_bounds(bounds: &[Point<R, D>]) -> Option<Self> {
-		let length = bounds.len().checked_sub(1).filter(|&length| length <= D)?;
-		let points = SMatrix::<R, D, D>::from_fn(|row, column| {
+	fn with_bounds(bounds: &[OPoint<R, D>]) -> Option<Self> {
+		let length = bounds.len().checked_sub(1).filter(|&len| len <= D::USIZE)?;
+		let points = OMatrix::<R, D, D>::from_fn(|row, column| {
 			if column < length {
 				bounds[column + 1].coords[row].clone() - bounds[0].coords[row].clone()
 			} else {
 				R::zero()
 			}
 		});
-		let points = points.view((0, 0), (D, length));
-		let matrix = SMatrix::<R, D, D>::from_fn(|row, column| {
+		let points = points.view((0, 0), (D::USIZE, length));
+		let matrix = OMatrix::<R, D, D>::from_fn(|row, column| {
 			if row < length && column < length {
 				points.column(row).dot(&points.column(column)) * (R::one() + R::one())
 			} else {
@@ -39,7 +57,7 @@ impl<R: RealField, const D: usize> Enclosing<R, D> for Ball<R, D> {
 			}
 		});
 		let matrix = matrix.view((0, 0), (length, length));
-		let vector = SVector::<R, D>::from_fn(|row, _column| {
+		let vector = OVector::<R, D>::from_fn(|row, _column| {
 			if row < length {
 				points.column(row).norm_squared()
 			} else {
@@ -49,11 +67,11 @@ impl<R: RealField, const D: usize> Enclosing<R, D> for Ball<R, D> {
 		let vector = vector.view((0, 0), (length, 1));
 		matrix.try_inverse().map(|matrix| {
 			let vector = matrix * vector;
-			let mut center = SVector::<R, D>::zeros();
+			let mut center = OVector::<R, D>::zeros();
 			for point in 0..length {
 				center += points.column(point) * vector[point].clone();
 			}
-			Ball {
+			Self {
 				center: &bounds[0] + &center,
 				radius_squared: center.norm_squared(),
 			}