+++ RELEASES.md
@@ -1,3 +1,8 @@
+# Version 0.3.0 (2024-03-17)
+
+  * Replace `Point` with `OPoint` supporting static as well as dynamic dimensions.
+  * Replace `ArrayVec<T, { D + 1 }>` with `OVec<T, DimNameSum<D, U1>>` supporting stabe Rust.
+
 # Version 0.2.0 (2023-04-01)
 
   * Update dependencies.